chore: container view 관리
